R6 Class Representing Tasks and Projects

A calendario object tracks a collection of projects that are contain tasks that are to be performed at different times

Details

BLAH BLAH BLAH

Methods

Public methods

Calendario$get_tasks()

Calendario$set_project()

Calendario$set_defaults()

Calendario$get_project()

Calendario$get_defaults()

Calendario$get_options()

Calendario$add_task()

Calendario$show_tasks()

Calendario$get_workload()

Calendario$get_calendar()

Calendario$show_calendar()

Calendario$clone()

Method get_tasks()

Retrieve a table of stored tasks

Usage

Calendario$get_tasks(...)

Arguments

...

Arguments passed to dplyr::filter()

Returns

A tibble with one row per task

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 22")
cal$get_tasks()

Method set_project()

Set the default project (special case of set_default())

Usage

Calendario$set_project(value)

Arguments

value

Name of the new default project

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$get_project()

Method set_defaults()

Set the default value for one or more specified task fields

Usage

Calendario$set_defaults(...)

Arguments

...

Name-value pairs specifying task fields and new default values

Examples

cal <- Calendario$new()
cal$set_defaults(type = "fun", team = "danielle")
cal$get_defaults()

Method get_project()

Retrieve the name of the current default project

Usage

Calendario$get_project()

Returns

A character vector of length 1

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$get_project()

Method get_defaults()

Retrieve the current defaults for all task fields

Usage

Calendario$get_defaults()

Returns

A tibble with one row

Examples

cal <- Calendario$new()
cal$get_defaults()

Method get_options()

Retrieve the options list

Usage

Calendario$get_options()

Returns

A list

Examples

cal <- Calendario$new()
cal$get_options()

Method add_task()

Add a task to a project

Usage

Calendario$add_task(
  description = NULL,
  start = NULL,
  stop = NULL,
  hours = NULL,
  type = NULL,
  project = NULL,
  team = NULL
)

Arguments

description

Character string providing a description of the task

start

Date the work starts (defaults to current date)

stop

Date the work stops (defaults to same day as start)

hours

Number of hours per day the task takes

type

Character string assigning the task to a category

project

Character string naming the project the task falls within

team

Character string describing the team

Examples

cal <- Calendario$new()
cal$add_task(
  description = "label for the task",
  start = as.Date("2025-01-24"),
  stop = as.Date("2025-01-26"),
  hours = 1,
  project = "my project",
  team = "just me"
)
cal$get_tasks()

Method show_tasks()

Show a table of stored tasks as a flextable

Usage

Calendario$show_tasks(...)

Arguments

...

Arguments passed to dplyr::filter()

Returns

A flextable object

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$show_tasks()

Method get_workload()

Retrieve a table describing daily workload

Usage

Calendario$get_workload(start = NULL, stop = NULL)

Arguments

start

Date the tabulation starts (defaults to current date)

stop

Date the tabulation stops (defaults to start date plus 90 days)

Returns

A tibble with one row per day

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$get_workload()

Method get_calendar()

Retrieve a list of tables describing monthly workload

Usage

Calendario$get_calendar(start = NULL, stop = NULL)

Arguments

start

Date the tabulation starts (defaults to current date)

stop

Date the tabulation stops (defaults to start date plus 90 days)

Returns

A tibble with one row per month

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$get_workload(start = "jan 1", stop = "feb 16")

Method show_calendar()

Show monthly workload as a flextable

Usage

Calendario$show_calendar(start = NULL, stop = NULL)

Arguments

start

Date the tabulation starts (defaults to current date)

stop

Date the tabulation stops (defaults to start date plus 90 days)

Returns

A flextable object

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$show_calendar(start = "jan 1", stop = "feb 16")

Method clone()

The objects of this class are cloneable with this method.

Usage

Calendario$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Calendario$get_tasks`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 22")
cal$get_tasks()
#> # A tibble: 2 × 7
#>   project    type  description start      stop       hours team 
#>   <chr>      <chr> <chr>       <date>     <date>     <dbl> <chr>
#> 1 my project NA    first task  2026-01-21 2025-06-08     1 NA   
#> 2 my project NA    second task 2026-01-22 2025-06-08     1 NA   

## ------------------------------------------------
## Method `Calendario$set_project`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$get_project()
#> [1] "my project"

## ------------------------------------------------
## Method `Calendario$set_defaults`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_defaults(type = "fun", team = "danielle")
cal$get_defaults()
#> # A tibble: 1 × 7
#>   project type  description start  stop   hours team    
#>   <chr>   <chr> <chr>       <date> <date> <dbl> <chr>   
#> 1 Project fun   Description NA     NA         1 danielle

## ------------------------------------------------
## Method `Calendario$get_project`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$get_project()
#> [1] "my project"

## ------------------------------------------------
## Method `Calendario$get_defaults`
## ------------------------------------------------

cal <- Calendario$new()
cal$get_defaults()
#> # A tibble: 1 × 7
#>   project type  description start  stop   hours team 
#>   <chr>   <chr> <chr>       <date> <date> <dbl> <chr>
#> 1 Project NA    Description NA     NA         1 NA   

## ------------------------------------------------
## Method `Calendario$get_options`
## ------------------------------------------------

cal <- Calendario$new()
cal$get_options()
#> $table_date_format
#> [1] "%a %b %d %Y"
#> 
#> $table_background
#> [1] "#ffffff"
#> 
#> $table_font_size
#> [1] 8
#> 
#> $table_theme
#> function (x) 
#> {
#>     if (!inherits(x, "flextable")) {
#>         stop(sprintf("Function `%s` supports only flextable objects.", 
#>             "theme_alafoli()"))
#>     }
#>     fp_bdr <- fp_border(width = flextable_global$defaults$border.width, 
#>         color = flextable_global$defaults$border.color)
#>     x <- border_remove(x)
#>     x <- bg(x, bg = "transparent", part = "all")
#>     x <- color(x, color = "#666666", part = "all")
#>     x <- bold(x = x, bold = FALSE, part = "all")
#>     x <- italic(x = x, italic = FALSE, part = "all")
#>     x <- padding(x = x, padding = 3, part = "all")
#>     x <- align_text_col(x, align = "left", header = TRUE)
#>     x <- align_nottext_col(x, align = "right", header = TRUE)
#>     x <- hline_bottom(x, part = "header", border = fp_bdr)
#>     x <- hline_top(x, part = "body", border = fp_bdr)
#>     fix_border_issues(x)
#> }
#> <bytecode: 0x55e28c01fd30>
#> <environment: namespace:flextable>
#> 
#> $date_range_start
#> function () 
#> lubridate::today()
#> <bytecode: 0x55e28bfefdb8>
#> <environment: namespace:calendario>
#> 
#> $date_range_stop
#> function () 
#> lubridate::today() + 90
#> <bytecode: 0x55e28bff0020>
#> <environment: namespace:calendario>
#> 
#> $date_task_start
#> function () 
#> lubridate::today()
#> <bytecode: 0x55e28bff0d08>
#> <environment: namespace:calendario>
#> 
#> $date_task_stop
#> function () 
#> lubridate::today()
#> <bytecode: 0x55e28bff0fe0>
#> <environment: namespace:calendario>
#> 

## ------------------------------------------------
## Method `Calendario$add_task`
## ------------------------------------------------

cal <- Calendario$new()
cal$add_task(
  description = "label for the task",
  start = as.Date("2025-01-24"),
  stop = as.Date("2025-01-26"),
  hours = 1,
  project = "my project",
  team = "just me"
)
cal$get_tasks()
#> # A tibble: 1 × 7
#>   project    type  description        start      stop       hours team   
#>   <chr>      <chr> <chr>              <date>     <date>     <dbl> <chr>  
#> 1 my project NA    label for the task 2025-01-24 2025-01-26     1 just me

## ------------------------------------------------
## Method `Calendario$show_tasks`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$show_tasks()

Project

Type

Description

Start date

End date

Daily hours

Team

my project

first task

Wed Jan 21 2026

Sun Jun 08 2025

1

my project

second task

Fri Jan 23 2026

Sun Jun 08 2025

1

## ------------------------------------------------
## Method `Calendario$get_workload`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$get_workload()
#> # A tibble: 165 × 6
#>    date       hours weekday month monthday  week
#>    <date>     <dbl> <ord>   <ord>    <int> <int>
#>  1 2025-06-09     2 Mon     Jun          9     1
#>  2 2025-06-10     2 Tue     Jun         10     1
#>  3 2025-06-11     2 Wed     Jun         11     1
#>  4 2025-06-12     2 Thu     Jun         12     1
#>  5 2025-06-13     2 Fri     Jun         13     1
#>  6 2025-06-16     2 Mon     Jun         16     2
#>  7 2025-06-17     2 Tue     Jun         17     2
#>  8 2025-06-18     2 Wed     Jun         18     2
#>  9 2025-06-19     2 Thu     Jun         19     2
#> 10 2025-06-20     2 Fri     Jun         20     2
#> # ℹ 155 more rows

## ------------------------------------------------
## Method `Calendario$get_calendar`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$get_workload(start = "jan 1", stop = "feb 16")
#> # A tibble: 181 × 6
#>    date       hours weekday month monthday  week
#>    <date>     <dbl> <ord>   <ord>    <int> <int>
#>  1 2025-06-09     2 Mon     Jun          9     1
#>  2 2025-06-10     2 Tue     Jun         10     1
#>  3 2025-06-11     2 Wed     Jun         11     1
#>  4 2025-06-12     2 Thu     Jun         12     1
#>  5 2025-06-13     2 Fri     Jun         13     1
#>  6 2025-06-16     2 Mon     Jun         16     2
#>  7 2025-06-17     2 Tue     Jun         17     2
#>  8 2025-06-18     2 Wed     Jun         18     2
#>  9 2025-06-19     2 Thu     Jun         19     2
#> 10 2025-06-20     2 Fri     Jun         20     2
#> # ℹ 171 more rows
    

## ------------------------------------------------
## Method `Calendario$show_calendar`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$show_calendar(start = "jan 1", stop = "feb 16")

Month

Days

Mon

Tue

Wed

Thu

Fri

Total

Jan

1-2

2

2

4

5-9

2

2

2

2

2

10

12-16

2

2

2

2

2

10

19-23

2

2

2

1

1

8

26-30

0

0

0

0

0

0

Feb

2-6

0

0

0

0

0

0

9-13

0

0

0

0

0

0

16

0

0

Jun

9-13

2

2

2

2

2

10

16-20

2

2

2

2

2

10

23-27

2

2

2

2

2

10

30

2

2

Jul

1-4

2

2

2

2

8

7-11

2

2

2

2

2

10

14-18

2

2

2

2

2

10

21-25

2

2

2

2

2

10

28-31

2

2

2

2

8

Aug

1

2

2

4-8

2

2

2

2

2

10

11-15

2

2

2

2

2

10

18-22

2

2

2

2

2

10

25-29

2

2

2

2

2

10

Sep

1-5

2

2

2

2

2

10

8-12

2

2

2

2

2

10

15-19

2

2

2

2

2

10

22-26

2

2

2

2

2

10

29-30

2

2

4

Oct

1-3

2

2

2

6

6-10

2

2

2

2

2

10

13-17

2

2

2

2

2

10

20-24

2

2

2

2

2

10

27-31

2

2

2

2

2

10

Nov

3-7

2

2

2

2

2

10

10-14

2

2

2

2

2

10

17-21

2

2

2

2

2

10

24-28

2

2

2

2

2

10

Dec

1-5

2

2

2

2

2

10

8-12

2

2

2

2

2

10

15-19

2

2

2

2

2

10

22-26

2

2

2

2

2

10

29-31

2

2

2

6