R6 Class Representing Tasks and Projects

A calendario object tracks a collection of projects that are contain tasks that are to be performed at different times

Details

BLAH BLAH BLAH

Methods

Public methods

Calendario$get_tasks()

Calendario$set_project()

Calendario$get_project()

Calendario$add_task()

Calendario$show_tasks()

Calendario$get_workload()

Calendario$get_calendar()

Calendario$show_calendar()

Calendario$clone()

Method get_tasks()

Retrieve a table of stored tasks

Usage

Calendario$get_tasks(...)

Arguments

...

Arguments passed to dplyr::filter()

Returns

A tibble with one row per task

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 22")
cal$get_tasks()

Method set_project()

Set the current project

Usage

Calendario$set_project(project)

Arguments

project

Name of the current project

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$get_project()

Method get_project()

Retrieve the name of the current project

Usage

Calendario$get_project()

Returns

A character vector of length 1

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$get_project()

Method add_task()

Add a task to a project

Usage

Calendario$add_task(
  description = NULL,
  start = NULL,
  stop = NULL,
  hours = NULL,
  type = NULL,
  project = NULL,
  team = NULL
)

Arguments

description

Character string providing a description of the task

start

Date the work starts (defaults to current date)

stop

Date the work stops (defaults to same day as start)

hours

Number of hours per day the task takes

type

Character string assigning the task to a category

project

Character string naming the project the task falls within

team

Character string describing the team

Examples

cal <- Calendario$new()
cal$add_task(
  description = "label for the task",
  start = as.Date("2025-01-24"),
  stop = as.Date("2025-01-26"),
  hours = 1,
  project = "my project",
  team = "just me"
)
cal$get_tasks()

Method show_tasks()

Show a table of stored tasks as a flextable

Usage

Calendario$show_tasks(...)

Arguments

...

Arguments passed to dplyr::filter()

Returns

A flextable object

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$show_tasks()

Method get_workload()

Retrieve a table describing daily workload

Usage

Calendario$get_workload(start = lubridate::today(), stop = start + 90)

Arguments

start

Date the tabulation starts (defaults to current date)

stop

Date the tabulation stops (defaults to start date plus 90 days)

Returns

A tibble with one row per day

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$get_workload()

Method get_calendar()

Retrieve a list of tables describing monthly workload

Usage

Calendario$get_calendar(start = lubridate::today(), stop = start + 90)

Arguments

start

Date the tabulation starts (defaults to current date)

stop

Date the tabulation stops (defaults to start date plus 90 days)

Returns

A tibble with one row per month

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$get_workload(start = "jan 1", stop = "feb 16")

Method show_calendar()

Show monthly workload as a flextable

Usage

Calendario$show_calendar(start = lubridate::today(), stop = start + 90)

Arguments

start

Date the tabulation starts (defaults to current date)

stop

Date the tabulation stops (defaults to start date plus 90 days)

Returns

A flextable object

Examples

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$show_calendar(start = "jan 1", stop = "feb 16")

Method clone()

The objects of this class are cloneable with this method.

Usage

Calendario$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Calendario$get_tasks`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 22")
cal$get_tasks()
#> # A tibble: 2 × 7
#>   project    type  description start      stop       hours team    
#>   <chr>      <chr> <chr>       <date>     <date>     <dbl> <chr>   
#> 1 my project work  first task  2026-01-21 2026-01-21     2 danielle
#> 2 my project work  second task 2026-01-22 2026-01-22     2 danielle

## ------------------------------------------------
## Method `Calendario$set_project`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$get_project()
#> [1] "my project"

## ------------------------------------------------
## Method `Calendario$get_project`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$get_project()
#> [1] "my project"

## ------------------------------------------------
## Method `Calendario$add_task`
## ------------------------------------------------

cal <- Calendario$new()
cal$add_task(
  description = "label for the task",
  start = as.Date("2025-01-24"),
  stop = as.Date("2025-01-26"),
  hours = 1,
  project = "my project",
  team = "just me"
)
cal$get_tasks()
#> # A tibble: 1 × 7
#>   project    type  description        start      stop       hours team   
#>   <chr>      <chr> <chr>              <date>     <date>     <dbl> <chr>  
#> 1 my project work  label for the task 2025-01-24 2025-01-26     1 just me

## ------------------------------------------------
## Method `Calendario$show_tasks`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$show_tasks()

project

type

description

start

stop

hours

team

my project

work

first task

Wednesday January 21 2026

Wednesday January 21 2026

2

danielle

my project

work

second task

Friday January 23 2026

Friday January 23 2026

2

danielle

## ------------------------------------------------
## Method `Calendario$get_workload`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$get_workload()
#> # A tibble: 67 × 6
#>    date       hours weekday month monthday  week
#>    <date>     <dbl> <ord>   <ord>    <int> <int>
#>  1 2025-06-02     0 Mon     Jun          2     1
#>  2 2025-06-03     0 Tue     Jun          3     1
#>  3 2025-06-04     0 Wed     Jun          4     1
#>  4 2025-06-05     0 Thu     Jun          5     1
#>  5 2025-06-06     0 Fri     Jun          6     1
#>  6 2025-06-09     0 Mon     Jun          9     2
#>  7 2025-06-10     0 Tue     Jun         10     2
#>  8 2025-06-11     0 Wed     Jun         11     2
#>  9 2025-06-12     0 Thu     Jun         12     2
#> 10 2025-06-13     0 Fri     Jun         13     2
#> # ℹ 57 more rows

## ------------------------------------------------
## Method `Calendario$get_calendar`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$get_workload(start = "jan 1", stop = "feb 16")
#> # A tibble: 33 × 6
#>    date       hours weekday month monthday  week
#>    <date>     <dbl> <ord>   <ord>    <int> <int>
#>  1 2026-01-01     0 Thu     Jan          1     0
#>  2 2026-01-02     0 Fri     Jan          2     0
#>  3 2026-01-05     0 Mon     Jan          5     1
#>  4 2026-01-06     0 Tue     Jan          6     1
#>  5 2026-01-07     0 Wed     Jan          7     1
#>  6 2026-01-08     0 Thu     Jan          8     1
#>  7 2026-01-09     0 Fri     Jan          9     1
#>  8 2026-01-12     0 Mon     Jan         12     2
#>  9 2026-01-13     0 Tue     Jan         13     2
#> 10 2026-01-14     0 Wed     Jan         14     2
#> # ℹ 23 more rows
    

## ------------------------------------------------
## Method `Calendario$show_calendar`
## ------------------------------------------------

cal <- Calendario$new()
cal$set_project("my project")
cal$add_task("first task", "jan 21")
cal$add_task("second task", "jan 23")
cal$show_calendar(start = "jan 1", stop = "feb 16")

Month

Days

Mon

Tue

Wed

Thu

Fri

Jan

1-2

0

0

5-9

0

0

0

0

0

12-16

0

0

0

0

0

19-23

0

0

2

0

2

26-30

0

0

0

0

0

Feb

2-6

0

0

0

0

0

9-13

0

0

0

0

0

16

0