Convenience functions

cal_task() and cal_project() are convenience functions that allow a piped workflow with calendario objects

cal_task() and cal_project() are convenience functions that allow a piped workflow with calendario objects

Usage

cal_add_task(x, ...)

cal_set_project(x, ...)

cal_get_workload(x, ...)

cal_get_calendar(x, ...)

cal_add_task(x, ...)

cal_get_tasks(x, ...)

cal_set_project(x, ...)

cal_get_workload(x, ...)

cal_get_calendar(x, ...)

Arguments

x

A calendario object

...

Arguments to be passed to the relevant R6 method

Value

Invisibly returns the calendario object

Invisibly returns the calendario object

Details

These functions are wrappers around one of the calendario object methods: cal_add_task() is a wrapper around $add_task(), cal_set_project() is a wrapper around $set_project() and so on. In all cases the functions take a calendario object x as their first argument, and all other arguments are passed via ... to the relevant calendario R6 method.

These functions are wrappers around one of the calendario object methods: cal_add_task() is a wrapper around $add_task(), cal_set_project() is a wrapper around $set_project() and so on. In all cases the functions take a calendario object x as their first argument, and all other arguments are passed via ... to the relevant calendario R6 method.

Examples

cal <- calendario$new()
cal |>
  cal_set_project("project-name") |>
  cal_add_task("description-1", "29 may") |>
  cal_add_task("description-2", "1 jun") |>
  cal_get_workload()
cal <- calendario$new()
cal |>
  cal_set_project("project-name") |>
  cal_add_task("description-1", "29 may") |>
  cal_add_task("description-2", "1 jun") |>
  cal_get_workload()
cal$get_tasks()
#> # A tibble: 2 × 7
#>   project      type  description   start      stop       hours team    
#>   <chr>        <chr> <chr>         <date>     <date>     <dbl> <chr>   
#> 1 project-name work  description-1 2025-05-29 2025-05-29     2 danielle
#> 2 project-name work  description-2 2025-06-01 2025-06-01     2 danielle